Car reservation, fast car reservation, grading rentacar service, cancelling reservation, adding, removing and editing branches done

+++ spring.project.isamrs/src/main/java/tim3/spring/project/isamrs/controller/CarController.java
@@ -1,5 +1,7 @@
 package tim3.spring.project.isamrs.controller;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -18,7 +20,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import tim3.spring.project.isamrs.dto.CarDTO;
+import tim3.spring.project.isamrs.dto.MessageDTO;
+import tim3.spring.project.isamrs.model.BranchOffice;
 import tim3.spring.project.isamrs.model.Car;
+import tim3.spring.project.isamrs.model.CarReservation;
+import tim3.spring.project.isamrs.model.CarReservationCriteriums;
 import tim3.spring.project.isamrs.model.Rentacar;
 import tim3.spring.project.isamrs.model.RentacarAdmin;
 import tim3.spring.project.isamrs.service.CarService;
@@ -36,7 +42,20 @@ public class CarController {
 
 	@Autowired
 	private CustomUserDetailsService userDetailsService;
+	
+	
 
+	@GetMapping(value = "/gradeCar/{id}/{grade}", produces = MediaType.APPLICATION_JSON_VALUE)
+	@PreAuthorize("hasRole('ROLE_USER')")
+	public ResponseEntity<Car> create(@PathVariable Long id, @PathVariable Integer grade) {
+		System.out.println("Uslo u grade car");
+		Car car = carService.getOne(id);
+		car.setScore(car.getScore()+grade);
+		car.setNumber(car.getNumber()+1);
+		carService.save(car);
+		return new ResponseEntity<>(car, HttpStatus.CREATED);
+	}
+	
 	@GetMapping(value = "/findConcreteCars/{rentacarId}")
 	public ResponseEntity<Set<Car>> findConcreteCars(@PathVariable String rentacarId) {
 		Rentacar retVal = rentacarService.getOne(Long.parseLong(rentacarId));
@@ -52,6 +71,7 @@ public class CarController {
 	@PostMapping(value = "/createCar", consumes = MediaType.APPLICATION_JSON_VALUE)
 	@PreAuthorize("hasRole('ROLE_RENTACAR_ADMIN')")
 	public ResponseEntity<Car> create(@RequestBody CarDTO carDTO) {
+		System.out.println("Uslo u car dto");
 		RentacarAdmin user = (RentacarAdmin) this.userDetailsService
 				.loadUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
 		Car newCar = new Car(carDTO);
@@ -73,16 +93,31 @@ public class CarController {
 
 	@DeleteMapping(value = "/deleteCar/{carId}")
 	@PreAuthorize("hasRole('ROLE_RENTACAR_ADMIN')")
-	public ResponseEntity<Car> deleteCar(@PathVariable String carId) {
+	public ResponseEntity<?> deleteCar(@PathVariable String carId) {
 		RentacarAdmin ra = (RentacarAdmin) this.userDetailsService
 				.loadUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
 		Car car = carService.getOne(Long.parseLong(carId));
 		if (car == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
-		ra.getRentacar().getCars().remove(car);
-		carService.delete(Long.parseLong(carId));
-		return new ResponseEntity<>(car, HttpStatus.OK);
+		Date now = new Date();
+		Boolean dozvola=true;
+		for (CarReservation r : car.getReservations()) {
+			if (r.getEndDate().compareTo(now) > 0) {
+				dozvola=false;
+			}
+		}
+		if (dozvola==true) {
+			ra.getRentacar().getCars().remove(car);
+			carService.delete(Long.parseLong(carId));
+			return new ResponseEntity<>(car, HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(new MessageDTO("Car is reserved, it cannot be deleted.", "Error"),
+					HttpStatus.OK);
+		}
+		
+		
+		
 	}
 
 	@GetMapping(value = "/findCar/{carId}")
@@ -93,6 +128,183 @@ public class CarController {
 		} else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
+	}
+	@SuppressWarnings("deprecation")
+	@GetMapping(value = "/findSuitCars/{rentacarId}/{startDate}/{endDate}/{startCity}/{endCity}/{carType}/{passengers}/{fromPrice}/{toPrice}", consumes=MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<List<Car>> findSuitCars(@PathVariable Long rentacarId,@PathVariable String startDate,
+			@PathVariable String endDate,@PathVariable String startCity,@PathVariable String endCity,
+			@PathVariable String carType,@PathVariable Integer passengers,@PathVariable Double fromPrice,@PathVariable Double toPrice)
+			
+			 {
+		List<Car> theFinalList=new ArrayList<Car>();
+		List<Car> finalList=new ArrayList<Car>();
+		List<Car> lista1 = new ArrayList<Car>();
+		List<Car> lista2 = new ArrayList<Car>();
+		Date startDatee = new Date(Integer.parseInt(startDate.split("\\-")[0]) - 1900,
+				Integer.parseInt(startDate.split("\\-")[1]) - 1, Integer.parseInt(startDate.split("\\-")[2]));
+		Date endDatee = new Date(Integer.parseInt(endDate.split("\\-")[0]) - 1900,
+				Integer.parseInt(endDate.split("\\-")[1]) - 1, Integer.parseInt(endDate.split("\\-")[2]));
+		
+		
+		
+		System.out.println("RENTACAR ID"+rentacarId);
+		Rentacar rentacar = rentacarService.getOne(rentacarId);
+		List<Car> cars = carService.findByRentacar(rentacar);
+		
+		Boolean mesto1=false;
+		Boolean mesto2=false;
+		for (BranchOffice b : rentacar.getBranches()) {
+			if (b.getCity().equalsIgnoreCase(startCity)) {
+				mesto1=true;
+			}
+			if (b.getCity().equalsIgnoreCase(endCity)) {
+				mesto2=true;
+			}
+		}
+		
+		
+		
+		if (mesto1==true && mesto2==true) {
+			System.out.println("Uslo u glavni if");
+			for (Car c : cars) {
+				if (c.getCarType().equalsIgnoreCase(carType)&&c.getSeats()>=passengers) {
+					lista1.add(c);
+				}
+			}
+			
+			System.out.println("Lista 1 size: "+lista1.size());
+		
+		
+			if (fromPrice!=-1) {
+				
+				for (Car c : lista1) {
+					if (c.getPrice()>=fromPrice) {
+						lista2.add(c);
+					}
+				}
+			}else {
+				for (Car c : lista1) {
+					lista2.add(c);
+				}
+			}
+			System.out.println("Lista 2 size: "+lista2.size());
+			
+			if (toPrice!=-1) {
+				for (Car c : lista2) {
+					if (c.getPrice()<=toPrice) {
+						finalList.add(c);
+					}
+				}
+			}else {
+				for (Car c : lista2) {
+					finalList.add(c);
+				}
+			}
+			System.out.println("Final lista size: "+finalList.size());
+			
+			for (Car car : finalList) {
+				Boolean dozvola=true;
+				for (CarReservation res : car.getReservations()) {
+					if (startDatee.compareTo(res.getStartDate()) < 0 && endDatee.compareTo(res.getEndDate()) < 0 && endDatee.compareTo(res.getStartDate()) > 0) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) <= 0 && endDatee.compareTo(res.getEndDate()) >= 0 ) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) >= 0 && endDatee.compareTo(res.getEndDate()) >= 0 && startDatee.compareTo(res.getEndDate()) < 0) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) >= 0 && endDatee.compareTo(res.getEndDate()) <= 0 ) {
+						dozvola=false;
+					}else {
+						dozvola=true;
+					}
+				}
+				if (dozvola==true) {
+					theFinalList.add(car);
+				}
+			}
+		}
+		
+		int brojac=-1;
+		  for (int i = 0 ; i < theFinalList.size();i++) { 
+			  brojac++;
+			  if (theFinalList.get(i).getOnFastRes()==true) {
+				  theFinalList.remove(brojac); 
+		  } 
+			  }
+			
+		System.out.println("Uslo u find suitable cars");
+	
+		return new ResponseEntity<>(theFinalList, HttpStatus.OK);
+		
+	}
+	
+	@SuppressWarnings("deprecation")
+	@GetMapping(value = "/findSuitCarsFast/{rentacarId}/{startDate}/{endDate}")
+	@PreAuthorize("hasRole('ROLE_USER')")
+	public ResponseEntity<List<Car>> findSuitCarsFast(@PathVariable String rentacarId,@PathVariable String startDate,
+			@PathVariable String endDate){
+		System.out.println("uSLO U FIND SUITABLE CARS FAST");
+			
+		List<Car> theFinalList=new ArrayList<Car>();
+		
+		Date startDatee = new Date(Integer.parseInt(startDate.split("\\-")[0]) - 1900,
+				Integer.parseInt(startDate.split("\\-")[1]) - 1, Integer.parseInt(startDate.split("\\-")[2]));
+		Date endDatee = new Date(Integer.parseInt(endDate.split("\\-")[0]) - 1900,
+				Integer.parseInt(endDate.split("\\-")[1]) - 1, Integer.parseInt(endDate.split("\\-")[2]));
+		
+		
+		
+		System.out.println("RENTACAR ID"+rentacarId);
+		Rentacar rentacar = rentacarService.getOne(Long.parseLong(rentacarId));
+		System.out.println("RENTACAR NAME "+rentacar.getName());
+		List<Car> cars = carService.findByRentacar(rentacar);
+		System.out.println("Preslo get cars");
+		for (Car car : cars) {
+			Boolean dozvola=true;
+			for (CarReservation res : car.getReservations()) {
+				if (endDate!="-1") {
+					if (startDatee.compareTo(res.getStartDate()) < 0 && endDatee.compareTo(res.getEndDate()) < 0 && endDatee.compareTo(res.getStartDate()) > 0) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) <= 0 && endDatee.compareTo(res.getEndDate()) >= 0 ) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) >= 0 && endDatee.compareTo(res.getEndDate()) >= 0 && startDatee.compareTo(res.getEndDate()) < 0) {
+						dozvola=false;
+					}
+					else if (startDatee.compareTo(res.getStartDate()) >= 0 && endDatee.compareTo(res.getEndDate()) <= 0 ) {
+						dozvola=false;
+					}else {
+						dozvola=true;
+					}
+				}else {
+					if (startDatee.compareTo(res.getStartDate()) >= 0 &&  startDatee.compareTo(res.getEndDate()) < 0) {
+						dozvola=false;
+					}
+				}
+				
+			}
+			if (dozvola==true) {
+				theFinalList.add(car);
+			}
+		}
+		
+		
+		int brojac=-1;
+		  for (int i = 0 ; i < theFinalList.size();i++) { 
+			  brojac++;
+			  if (theFinalList.get(i).getOnFastRes()==false) {
+				  theFinalList.remove(brojac); 
+		  } 
+			  }
+		 
+			
+		
+		
+		return new ResponseEntity<>(theFinalList, HttpStatus.OK);
+		
 	}
 
 	@PutMapping(value = "/saveEditedCar", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -106,9 +318,15 @@ public class CarController {
 		c.setName(car.getName());
 		c.setPrice(car.getPrice());
 		c.setYear(car.getYear());
+		c.setCarType(car.getCarType());
+		c.setBrand(car.getBrand());
+		c.setModel(car.getModel());
+		c.setSeats(car.getSeats());
 
 		Car editedCar = carService.save(c);
 		return new ResponseEntity<>(editedCar, HttpStatus.OK);
 	}
+	
+	
 
 }
